+++ packages/x-components/src/x-modules/tagging/service/types.ts
@@ -5,14 +5,23 @@ import { Result } from '@empathyco/x-types';
  *
  * @public
  */
-export interface PDPAddToCartService {
+export interface ExternalTaggingService {
   /**
    * Stores in the local storage the information from the Result clicked by the user
-   * in order to be able to track later on.
+   * in order to be able to track the add to cart later on the result's PDP.
    *
    * @param result - The result to store.
    */
   storeResultClicked(result: Result): void;
+
+  /**
+   * Stores in the session storage the information from the Result added to the cart
+   * by the user in order to be able to track the checkout later on when the checkout
+   * process has been completed by shopper.
+   *
+   * @param result - The result to store.
+   */
+  storeAddToCart(result: Result): void;
 
   /**
    * Checks if the local storage contains a result information for the given id and moves
@@ -23,8 +32,9 @@ export interface PDPAddToCartService {
   moveToSessionStorage(id?: string): void;
 
   /**
-   * Checks if the session storage contains a result information for given id or the current url
-   * and tracks the add to cart if exists.
+   * Checks if the session storage contains a result information for a given id or the current url.
+   * If exists, it tracks the add to cart and saves the add to cart information into session
+   * storage.
    *
    * @param id - The id of the result to track.
    */